[fix] change unnecessary variable type
long long to int

+++ seoul_21_minjaekim/BoJ/Gold I/boj2014/boj2014.cpp
@@ -22,7 +22,7 @@ int main()
 
     for (int i = 1; i < n + 1; ++i)
     {
-        long long min_number = numeric_limits<long long>::max();
+        int min_number = numeric_limits<int>::max();
         int min_index[100];
         int min_count = 0;
         for (int j = 0; j < k; ++j)
